fix（ user ）: 🐛 表单不能修改头像 、查询部门参数类型错误

+++ yf-boot-admin/yf-base-model/src/main/java/com/yf/model/system/query/UserPageQuery.java
@@ -26,7 +26,7 @@ public class UserPageQuery extends BasePageQuery {
     private Integer status;
 
     @Schema(description = "部门ID")
-    private Long deptId;
+    private Integer deptId;
 
     @Schema(description = "创建时间-开始时间")
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
